# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Academic Knowledge Backend",
    "description": "Backend API for course-centric academic knowledge management.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Courses Endpoint",
        "operationId": "list_courses_endpoint_api_v1_courses_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": true,
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseRead"
                  },
                  "title": "Response List Courses Endpoint Api V1 Courses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Create Course Endpoint",
        "operationId": "create_course_endpoint_api_v1_courses_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{course_code}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course Endpoint",
        "operationId": "get_course_endpoint_api_v1_courses__course_code__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "courses"
        ],
        "summary": "Update Course Endpoint",
        "operationId": "update_course_endpoint_api_v1_courses__course_code__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{course_code}/seed": {
      "put": {
        "tags": [
          "courses"
        ],
        "summary": "Update Course Seed Endpoint",
        "operationId": "update_course_seed_endpoint_api_v1_courses__course_code__seed_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseSeedUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{course_code}/sla": {
      "put": {
        "tags": [
          "courses"
        ],
        "summary": "Set Sla Endpoint",
        "operationId": "set_sla_endpoint_api_v1_courses__course_code__sla_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{course_code}/documents": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Course Documents",
        "operationId": "list_course_documents_api_v1_courses__course_code__documents_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Code"
            }
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentTier"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "recent",
              "title": "Sort"
            }
          },
          {
            "name": "subtype",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subtype"
            }
          },
          {
            "name": "topic_tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{course_code}/topic-tags/regenerate": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Regenerate Topic Tags Endpoint",
        "operationId": "regenerate_topic_tags_endpoint_api_v1_courses__course_code__topic_tags_regenerate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicTagsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/leaderboard/course/{course_code}": {
      "get": {
        "tags": [
          "voting"
        ],
        "summary": "Get Course Leaderboard",
        "description": "Get top contributors for a specific course.",
        "operationId": "get_course_leaderboard_api_v1_leaderboard_course__course_code__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Code"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Course Leaderboard Api V1 Leaderboard Course  Course Code  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseCreate": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 20,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "topic_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Summary"
          },
          "short_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Description"
          },
          "review_sla_hours": {
            "type": "integer",
            "maximum": 72.0,
            "minimum": 24.0,
            "title": "Review Sla Hours",
            "default": 48
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "CourseCreate"
      },
      "CourseRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "topic_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Summary"
          },
          "short_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Description"
          },
          "topic_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topic Tags"
          },
          "review_sla_hours": {
            "type": "integer",
            "title": "Review Sla Hours"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "description",
          "topic_summary",
          "short_description",
          "topic_tags",
          "review_sla_hours",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "CourseRead"
      },
      "CourseSeedUpdate": {
        "properties": {
          "topic_summary": {
            "type": "string",
            "title": "Topic Summary"
          },
          "short_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Description"
          }
        },
        "type": "object",
        "required": [
          "topic_summary"
        ],
        "title": "CourseSeedUpdate"
      },
      "CourseUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "topic_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Summary"
          },
          "short_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Description"
          },
          "review_sla_hours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 72.0,
                "minimum": 24.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Sla Hours"
          }
        },
        "type": "object",
        "title": "CourseUpdate"
      },
...
      "DocumentListRead": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "DocumentListRead"
      },
      "DocumentRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "uploader_id": {
            "type": "string",
            "format": "uuid",
            "title": "Uploader Id"
          },
          "document_tier": {
            "type": "string",
            "title": "Document Tier"
          },
          "material_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Type"
          },
          "contribution_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution Type"
          },
          "topic_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topic Tags"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "is_active_version": {
            "type": "boolean",
            "title": "Is Active Version"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "file_format": {
            "type": "string",
            "title": "File Format"
          },
          "storage_raw_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Raw Path"
          },
          "storage_md_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Md Path"
          },
          "no_reviewer_flag": {
            "type": "boolean",
            "title": "No Reviewer Flag"
          },
          "sla_breached": {
            "type": "boolean",
            "title": "Sla Breached"
          },
          "sla_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Deadline"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "uploader_id",
          "document_tier",
          "material_type",
          "contribution_type",
          "topic_tags",
          "status",
          "version",
          "is_active_version",
          "original_filename",
          "file_format",
          "storage_raw_path",
          "storage_md_path",
          "no_reviewer_flag",
          "sla_breached",
          "sla_deadline",
          "uploaded_at",
          "updated_at"
        ],
        "title": "DocumentRead"
      },
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "UPLOADED",
          "PARSING",
          "EVALUATING",
          "NEEDS_REVIEW",
          "APPROVED",
          "INDEXING",
          "INDEXED",
          "REJECTED",
          "FAILED"
        ],
        "title": "DocumentStatus"
      },
      "DocumentTier": {
        "type": "string",
        "enum": [
          "official",
          "community"
        ],
        "title": "DocumentTier"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SlaUpdate": {
        "properties": {
          "sla_hours": {
            "type": "integer",
            "maximum": 72.0,
            "minimum": 24.0,
            "title": "Sla Hours"
          }
        },
        "type": "object",
        "required": [
          "sla_hours"
        ],
        "title": "SlaUpdate"
      },
...
      "TopicTagsRead": {
        "properties": {
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "tags"
        ],
        "title": "TopicTagsRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}